feat: restore round filter dropdown
