+++ packages/grant-explorer/src/features/discovery/FilterDropdown.tsx
@@ -1,71 +1,120 @@
-import { Disclosure } from "@headlessui/react";
+import { Fragment } from "react";
+import { Link, useSearchParams } from "react-router-dom";
+import { Transition, Disclosure } from "@headlessui/react";
 import { ChevronDownIcon, ChevronUpIcon } from "@heroicons/react/20/solid";
-import isEqual from "lodash-es/isEqual";
 
 import { Dropdown, DropdownItem } from "../common/Dropdown";
+import { toQueryString } from "./RoundsFilter";
+import { FilterStatus, parseFilterParams } from "./hooks/useFilterRounds";
+import { ROUND_PAYOUT_DIRECT, ROUND_PAYOUT_MERKLE } from "common";
+import { getFilterLabel } from "./utils/getFilterLabel";
+import { allChains } from "../../app/chainConfig";
 
-export type FilterDropdownOption<T> = {
+export type FilterOption = {
   label: string;
-  children: { label: string; value: T }[];
-  allowMultiple: boolean;
+  value: string;
+  hide?: boolean;
+  children?: FilterOption[];
+};
+export const filterOptions: FilterOption[] = [
+  {
+    label: "All",
+    value: "",
+    children: [],
+  },
+  {
+    label: "Type",
+    value: "type",
+    children: [
+      {
+        label: "Quadratic funding",
+        value: ROUND_PAYOUT_MERKLE,
+      },
+      {
+        label: "Direct grants",
+        value: ROUND_PAYOUT_DIRECT,
+      },
+    ],
+  },
+
+  {
+    label: "Status",
+    value: "status",
+    children: [
+      {
+        label: "Active",
+        value: FilterStatus.active,
+      },
+      {
+        label: "Taking applications",
+        value: FilterStatus.taking_applications,
+      },
+      {
+        label: "Finished",
+        value: FilterStatus.finished,
+      },
+    ],
+  },
+  {
+    label: "Network",
+    value: "network",
+    children: allChains.map(({ id, name }) => ({
+      label: `Rounds on ${name}`,
+      value: String(id),
+    })),
+  },
+];
+
+export type FilterProps = {
+  type: string;
+  status: string;
+  network: string;
 };
 
-type FilterDropdownProps<T> = {
-  options: FilterDropdownOption<T>[];
-  onChange: (selected: T[]) => void;
-  selected: T[];
-};
+export function FilterDropdown() {
+  const [params] = useSearchParams();
 
-export function FilterDropdown<T>(props: FilterDropdownProps<T>): JSX.Element {
-  const options = props.options.flatMap((option) => {
-    return option.children;
-  });
+  const filter = parseFilterParams(params);
+  const { status = "", type = "", network = "" } = filter;
 
-  const selected = options.filter((option) =>
-    props.selected.some((selected) => isEqual(selected, option.value))
-  );
-
-  const selectedLabels = selected.map((option) => option.label);
-  const selectedValues = selected.map((option) => option.value);
-
-  let selectedLabel = "All";
-
-  if (selectedLabels.length === 1) {
-    selectedLabel = selectedLabels[0];
-  } else if (selectedLabels.length > 1) {
-    selectedLabel = "Multiple";
-  }
-
+  const selected = getFilterLabel({ status, type, network });
   return (
     <Dropdown
-      label={selectedLabel}
-      options={props.options}
+      label={selected?.label}
+      options={filterOptions}
       keepOpen
-      headerElement={(close) => (
-        <button
-          className="relative w-[340px] py-2 pl-2 pr-10 text-left hover:bg-grey-100"
-          onClick={() => {
-            props.onChange([]);
-            close();
-          }}
-        >
-          All
-        </button>
-      )}
-      renderItem={(item) => {
-        const anyChildrenSelected = item.children.some((child) =>
-          props.selected.some((value) => isEqual(value, child.value))
-        );
+      renderItem={({ label, value: filterKey, children, close }) => {
+        // Filters can be multi selected (ie many networks)
+        const selectedFilter =
+          { status, type, network }[filterKey]?.split(",").filter(Boolean) ??
+          [];
 
+        if (!children?.length) {
+          return (
+            <DropdownItem
+              $as={Link}
+              to={`/rounds?${toQueryString({
+                ...Object.fromEntries(params),
+                network: "",
+                status: "",
+                type: "",
+              })}`}
+              preventScrollReset
+              onClick={() => close()}
+            >
+              {label}
+            </DropdownItem>
+          );
+        }
         return (
-          <Disclosure defaultOpen={anyChildrenSelected}>
+          <Disclosure defaultOpen={Boolean(selectedFilter.length)}>
             <div className="relative mt-1">
-              <Disclosure.Button className="relative w-full py-2 pl-2 pr-10 text-left hover:bg-grey-100">
+              <Disclosure.Button className="relative w-[340px] py-2 pl-3 pr-10 text-left hover:bg-grey-100">
                 {({ open }) => {
                   const Icon = open ? ChevronUpIcon : ChevronDownIcon;
                   return (
                     <>
-                      <span className="block truncate">{item.label}</span>
+                      <span className="block truncate">{label}</span>
                       <span className="pointer-events-none absolute inset-y-0 right-0 flex items-center pr-2">
                         <Icon
                           className="h-5 w-5 text-gray-400"
@@ -76,66 +125,51 @@ export function FilterDropdown<T>(props: FilterDropdownProps<T>): JSX.Element {
                   );
                 }}
               </Disclosure.Button>
-              <Disclosure.Panel className="mt-1 w-full overflow-auto p-2">
-                {item.children.map((child, j) => {
-                  const isSelected = props.selected.some((value) =>
-                    isEqual(value, child.value)
-                  );
 
-                  const onTrigger = () => {
-                    let newSelected: T[] = [...selectedValues];
+              <Transition
+                as={Fragment}
+                leave="transition ease-in duration-100"
+                leaveFrom="opacity-100"
+                leaveTo="opacity-0"
+              >
+                <Disclosure.Panel
+                  static
+                  className=" mt-1 w-full overflow-auto p-2"
+                >
+                  {children
+                    ?.filter((child) => !child.hide)
+                    .map((child, j) => {
+                      const isChecked = selectedFilter?.includes(child.value);
 
-                    // clear sibling options if allowMultiple is false
-                    if (item.allowMultiple === false) {
-                      newSelected = newSelected.filter(
-                        (option) =>
-                          !item.children.some((child) =>
-                            isEqual(child.value, option)
-                          )
+                      const nextValue = isChecked
+                        ? // Remove or add the value
+                          selectedFilter?.filter((f) => f !== child.value)
+                        : selectedFilter?.concat(child.value);
+                      return (
+                        <DropdownItem
+                          key={j}
+                          $as={Link}
+                          to={`/rounds?${toQueryString({
+                            // Merge existing search params (so it doesn't reset sorting or the other selections)
+                            ...filter,
+                            [filterKey]: nextValue.join(","),
+                          })}`}
+                        >
+                          <div className="flex gap-2">
+                            <input type="checkbox" checked={isChecked} />
+                            <span
+                              className={`block truncate ${
+                                selected ? "font-medium" : "font-normal"
+                              }`}
+                            >
+                              {child.label}
+                            </span>
+                          </div>
+                        </DropdownItem>
                       );
-                    }
-
-                    if (!isSelected) {
-                      newSelected.push(child.value);
-                    } else {
-                      newSelected = newSelected.filter((option) => {
-                        return !isEqual(option, child.value);
-                      });
-                    }
-
-                    props.onChange(newSelected);
-                  };
-
-                  // HeadlessUI binds onclick so clicking the label doesn't work to trigger the checkbox
-                  return (
-                    <DropdownItem
-                      key={j}
-                      $as="label"
-                      onClick={(e) => {
-                        e.preventDefault();
-                        onTrigger();
-                      }}
-                      className={`block truncate ${
-                        isSelected ? "font-medium" : "font-normal"
-                      }`}
-                    >
-                      <input
-                        type="checkbox"
-                        className="mr-2"
-                        checked={isSelected}
-                        onChange={() => {
-                          // do nothing
-                        }}
-                        onClick={(e) => {
-                          e.stopPropagation();
-                          onTrigger();
-                        }}
-                      />
-                      {child.label}
-                    </DropdownItem>
-                  );
-                })}
-              </Disclosure.Panel>
+                    })}
+                </Disclosure.Panel>
+              </Transition>
             </div>
           </Disclosure>
         );
