+++ packages/grant-explorer/src/features/discovery/hooks/useFilterRounds.ts
@@ -5,7 +5,7 @@ import {
   getActiveChainIds,
   useRounds,
 } from "../../api/rounds";
-import { FilterProps } from "../RoundFilterDropdown";
+import { FilterProps } from "../FilterDropdown";
 import { SortProps } from "../SortDropdown";
 import { createRoundsStatusFilter } from "../utils/createRoundsStatusFilter";
 import { ROUND_PAYOUT_MERKLE } from "common";
