+++ packages/grant-explorer/src/features/discovery/utils/getExplorerPageTitle.ts
@@ -1,5 +1,5 @@
 import { ROUND_PAYOUT_DIRECT, ROUND_PAYOUT_MERKLE } from "common";
-import { FilterProps } from "../RoundFilterDropdown";
+import { FilterProps } from "../FilterDropdown";
 import { getFilterLabel } from "./getFilterLabel";
 import { FilterStatus } from "../hooks/useFilterRounds";
